fix long loading for external projects in financial transfer licensing service
