+++ src/app/modules/services/financial-transfer-licensing/pages/financial-transfers-licensing/financial-transfers-licensing.component.html
@@ -238,7 +238,7 @@
       <financial-transfers-projects #financialTransfersProjectsTab [list]="model?.financialTransfersProjects  || []"
         [readonly]="readonly" (listUpdated)="updateTransactionAmount($event)"
         [submissionMechanism]="submissionMechanism"
-        [approvedFinancialTransferProjects]="approvedFinancialTransferProjectsList"
+        
         [requestType]="model?.requestType || financialTransferRequestTypesEnum.NEW"
         (financialTransfersProjectListUpdated)="onProjectsListUpdated($event)">
       </financial-transfers-projects>
